==> code/01 - APA Attendence Geocoding.Rmd <==
---
title: "01 - APA Attendence Geocoding"
author: "Alec Stashevsky"
date: "June 21, 2020"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(data.table)
library(readxl)
library(openxlsx)
library(ggmap) # Need to cite 
rm(list = ls()) # Clean R Environment
```


##################################   NOTES   ############################
1. Need to make sure all CITY == NEW YORK is coded as NEW YORK CITY
2. Need to rerun NYC and SF Geocode with city fixes
3. Change read_xlsx to read_excel so it can be more compatible for users

* ASK JOSH ABOUT TRAVEL DISTANCE CALCULATION

## Import Attendence Data

```{r Import}
# Import the Attendance data
path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Data/Registrant Locations 2018-2020.xlsx"

nyc.2018.raw <- read_excel(path.in, sheet = "2018 NYC") %>% setDT()
sf.2019.raw <- read_excel(path.in, sheet = "2019 SF") %>% setDT()
phl.2020.raw <- read_excel(path.in, sheet = "2020 PHL") %>% setDT()

# Clean up unnecessary columns - total records summary
nyc.2018 <- nyc.2018.raw[,1:(length(nyc.2018.raw)-2)]
sf.2019 <- sf.2019.raw[,1:(length(sf.2019.raw)-2)]
phl.2020 <- phl.2020.raw[,1:(length(phl.2020.raw)-2)]
```

## Prep data for Google Geocoding API
```{r Geocode Prep}
# Convention Center Addresses
## Pennsylvania Convention center - Philadelphia
## Moscone Center - San Francisco
## Javitz Center - New York
convention.address <- c("1101 Arch St, Philadelphia, PA 19107",
             "747 Howard St, San Francisco, CA 94103",
             "429 11th Ave, New York, NY 10001")

# Fix "Not Applicable" values to true NAs
## Gets indexes
nyc.na <- which(nyc.2018$STATE == "NOT APPLICABLE", arr.ind = TRUE)
sf.na <- which(sf.2019$STATE == "NOT APPLICABLE", arr.ind = TRUE)
phl.na <- which(phl.2020$STATE == "NOT APPLICABLE", arr.ind = TRUE)

## Replace with NA
nyc.2018[nyc.na]$STATE <- NA
sf.2019[sf.na]$STATE <- NA
phl.2020[phl.na]$STATE <- NA

# Now lets begin cleaning minor issues that errors in Google API


# Correct Data entry error detected on initial Google API run
## Mostly city names that were misspelled

### NYC
nyc.2018[CITY == "IKAAN"]$CITY <- "Iksan"
nyc.2018[CITY == "Queron"]$CITY <- "Quezon"
nyc.2018[CITY == "Boeke"]$CITY <- "Boekel"
nyc.2018[CITY == "Not Applicable"]$CITY <- NA

### SF
sf.2019[CITY == "Godoy Cruz"]$CITY <- "Godoy Cruz"
sf.2019[CITY == "Burlingame"]$CITY <- "Burlingame"
sf.2019[CITY == "Ska"]$CITY <- "Skå"

## PHL is good to go!
```

```{r Geocode Location Set}
# Prep single location column to feed Google API
## Remove US country name due to Google API issues
nyc.prep <- nyc.2018[,
                     location :=
                       ifelse(is.na(STATE),
                              paste(CITY, COUNTRY, sep = ", "),
                              ifelse(COUNTRY == "United States",
                                     paste(CITY, STATE, "US", sep =", "),
                                     paste(CITY, STATE, COUNTRY, sep = ", ")))]

sf.prep <- sf.2019[, 
                     location :=
                       ifelse(is.na(STATE),
                              paste(CITY, COUNTRY, sep = ", "),
                              ifelse(COUNTRY == "United States",
                                     paste(CITY, STATE, "US", sep =", "),
                                     paste(CITY, STATE, COUNTRY, sep = ", ")))]
phl.prep <- phl.2020[, 
                     location :=
                       ifelse(is.na(STATE),
                              paste(CITY, COUNTRY, sep = ", "),
                              ifelse(COUNTRY == "United States",
                                     paste(CITY, STATE, "US", sep =", "),
                                     paste(CITY, STATE, COUNTRY, sep = ", ")))]

# There are some attendees who only have a country coded
nyc.fix <- which(nyc.prep$location %like% "^NA, ", arr.ind = TRUE)
sf.fix <- which(sf.prep$location %like% "^NA, ", arr.ind = TRUE)
phl.fix <- which(phl.prep$location %like% "^NA, ", arr.ind = TRUE)

# Overwrite location for entries with no city and state
nyc.prep[nyc.fix]$location <- nyc.prep[nyc.fix]$COUNTRY
sf.prep[sf.fix]$location <- sf.prep[sf.fix]$COUNTRY
phl.prep[phl.fix]$location <- phl.prep[phl.fix]$COUNTRY

# Check NA processing for correct location 
nyc.prep[is.na(CITY) & is.na(STATE)] # 1 Argentina, 1 Finland, 47 US
sf.prep[is.na(CITY) & is.na(STATE)] # 1 Brazil, Finland, Chile ; 51 US
phl.prep[is.na(CITY) & is.na(STATE)] #1 Canada, 1 Chile, 21 US

############################ QUESTIONS FOR JOSH ###############################

# We need to check how Google API handles this, may need to drop
# There are a lot of Area X Rep inputs to the city - are these APA districts??
# DISCUSS WITH JOSH
```

## Geocode
```{r Geocoding NYC}
# Geocode with Google Geocoding API - this will take some time
nyc.geo <- mutate_geocode(nyc.prep, location = location)

# Check errors
nyc.geo[is.na(lat) | is.na(lon)]
```

```{r Geocoding SF}
# Geocode with Google Geocoding API - this will take some time
sf.geo <- mutate_geocode(sf.prep, location = location)

# Check errors
view(sf.geo[is.na(lat) | is.na(lon)])
```

```{r Geocoding PHL}
# Geocode with Google Geocoding API - this will take some time
phl.geo <- mutate_geocode(phl.prep, location = location)

# Check errors
phl.geo[is.na(lat) | is.na(lon)]
```

```{r Convention Centers}
# Import list of convention centers
convention.import <- read_excel("C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/Convention Centers.xlsx", range = "A1:E15")

# Geocode conventions centers
conventions.geo <- mutate_geocode(convention.import, location = location)

# Export 
write.xlsx(conventions.geo, "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Convention Center GeoData.xlsx")
```

```{r Final Write}
geo.sets <- list("NYC 2018" = nyc.geo,
                 "SF 2019" = sf.geo,
                 "PHL 2020" = phl.geo,
                 "Convention Centers" = conventions.geo)

write.xlsx(geo.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA GeoData.xlsx")
```
