==> code/03 - APA Geodistance Calculator.Rmd <==
---
title: "03 - APA Geodistance Calculator"
author: "Alec Stashevsky"
date: "July 4, 2020"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(openxlsx)
library(tidyverse)
library(data.table)
library(geosphere)
rm(list = ls()) # Clean R Environment
```


## Import Geo-unique Data
```{r Import}
path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Geo-Unqiue Frequency Data.xlsx"

path.convention <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Convention Center GeoData.xlsx"

nyc.raw <- read_excel(path.in, sheet = "NYC 2018") %>% setDT()
sf.raw <- read_excel(path.in, sheet = "SF 2019") %>% setDT()
phl.raw <- read_excel(path.in, sheet = "PHL 2020") %>% setDT()
conventions <- read_xlsx(path.convention) %>%
  setDT()
```

## Clean Data
```{r Clean}
# We need to remove Military Post office locations
military_po <- c("APO", "AE", "AA", "FPO", "DPO", "MPO")

# Locate invalid locations - mostly people who wrote the APA Area or military
# Remove invalid coordinates
nyc <- nyc.raw[!is.na(lat) | !is.na(lon)][!(CITY %like% "Area")][
  !(CITY %in% military_po)]

sf <- sf.raw[!is.na(lat) | !is.na(lon)][!(CITY %like% "Area")][!(CITY %like% "AREA")]

# There are no issues with representative areas in PHL
phl <- phl.raw[!is.na(lat) | !is.na(lon)]

```


## Calculate Geodistance to Convention Center
```{r Geodistance}
# Get convention center coordinates
nyc.dest <- as.numeric(conventions[Name =="Javitz Center", c('lon', 'lat')])
sf.dest <- as.numeric(conventions[Name =="Moscone Center", c('lon', 'lat')])
phl.dest <- as.numeric(conventions[Name =="Pennsylvania Convention center",
                                   c('lon', 'lat')])

dests <- conventions[, .(lon, lat)]

# Calculate Geodistance in meters - with ellipsoidal method
for(i in 1:nrow(conventions)){
  assign(paste0("nyc.gdist.", conventions$CITY[i]),
        distGeo(nyc[, c('lon', 'lat')], dests[i]))
}

for(i in 1:nrow(conventions)){
  assign(paste0("sf.gdist.", conventions$CITY[i]),
        distGeo(sf[, c('lon', 'lat')], dests[i]))
}

for(i in 1:nrow(conventions)){
  assign(paste0("phl.gdist.", conventions$CITY[i]),
        distGeo(phl[, c('lon', 'lat')], dests[i]))
}
# Merge distances together
nyc.gdist.total <- as.data.table(cbind(gdist.PHL = nyc.gdist.Philadelphia,
                                       gdist.SFO = `nyc.gdist.San Francisco`,
                                       gdist.JFK = `nyc.gdist.New York City`,
                                       gdist.ORD = nyc.gdist.Chicago,
                                       gdist.MSY = `nyc.gdist.New Orleans`,
                                       gdist.YYZ = nyc.gdist.Toronto,
                                       gdist.LAX = `nyc.gdist.Los Angeles`,
                                       gdist.DFW = nyc.gdist.Dallas,
                                       gdist.DCA = `nyc.gdist.Washington DC`,
                                       gdist.YUL = nyc.gdist.Montreal,
                                       gdist.MIA = nyc.gdist.Miami,
                                       gdist.SAN = `nyc.gdist.San Diego`,
                                       gdist.ATL = nyc.gdist.Atlanta,
                                       gdist.HNL = nyc.gdist.Honolulu))

sf.gdist.total <- as.data.table(cbind(gdist.PHL = sf.gdist.Philadelphia,
                                      gdist.SFO = `sf.gdist.San Francisco`,
                                      gdist.JFK = `sf.gdist.New York City`,
                                      gdist.ORD = sf.gdist.Chicago,
                                      gdist.MSY = `sf.gdist.New Orleans`,
                                      gdist.YYZ = sf.gdist.Toronto,
                                      gdist.LAX = `sf.gdist.Los Angeles`,
                                      gdist.DFW = sf.gdist.Dallas,
                                      gdist.DCA = `sf.gdist.Washington DC`,
                                      gdist.YUL = sf.gdist.Montreal,
                                      gdist.MIA = sf.gdist.Miami,
                                      gdist.SAN = `sf.gdist.San Diego`,
                                      gdist.ATL = sf.gdist.Atlanta,
                                      gdist.HNL = sf.gdist.Honolulu))

phl.gdist.total <- as.data.table(cbind(gdist.PHL = phl.gdist.Philadelphia,
                                       gdist.SFO = `phl.gdist.San Francisco`,
                                       gdist.JFK = `phl.gdist.New York City`,
                                       gdist.ORD = phl.gdist.Chicago,
                                       gdist.MSY = `phl.gdist.New Orleans`,
                                       gdist.YYZ = phl.gdist.Toronto,
                                       gdist.LAX = `phl.gdist.Los Angeles`,
                                       gdist.DFW = phl.gdist.Dallas,
                                       gdist.DCA = `phl.gdist.Washington DC`,
                                       gdist.YUL = phl.gdist.Montreal,
                                       gdist.MIA = phl.gdist.Miami,
                                       gdist.SAN = `phl.gdist.San Diego`,
                                       gdist.ATL = phl.gdist.Atlanta,
                                       gdist.HNL = phl.gdist.Honolulu))

# Merge onto Geo-unique data
final.nyc <- cbind(nyc, nyc.gdist.total)
final.sf <- cbind(sf, sf.gdist.total)
final.phl <- cbind(phl, phl.gdist.total)

# Create indicator variable for driving if less than 400 km away from meeting
final.nyc[, `:=`(drive.PHL = ifelse(gdist.PHL <= 400000, 1, 0),
                 drive.SFO = ifelse(gdist.SFO <= 400000, 1, 0),
                 drive.JFK = ifelse(gdist.JFK <= 400000, 1, 0),
                 drive.ORD = ifelse(gdist.ORD <= 400000, 1, 0),
                 drive.MSY = ifelse(gdist.MSY <= 400000, 1, 0),
                 drive.YYZ = ifelse(gdist.YYZ <= 400000, 1, 0),
                 drive.LAX = ifelse(gdist.LAX <= 400000, 1, 0),
                 drive.DFW = ifelse(gdist.DFW <= 400000, 1, 0),
                 drive.DCA = ifelse(gdist.DCA <= 400000, 1, 0),
                 drive.YUL = ifelse(gdist.YUL <= 400000, 1, 0),
                 drive.MIA = ifelse(gdist.MIA <= 400000, 1, 0),
                 drive.SAN = ifelse(gdist.SAN <= 400000, 1, 0),
                 drive.ATL = ifelse(gdist.ATL <= 400000, 1, 0),
                 drive.HNL = ifelse(gdist.HNL <= 400000, 1, 0)
                 )]

final.sf[, `:=`(drive.PHL = ifelse(gdist.PHL <= 400000, 1, 0),
                drive.SFO = ifelse(gdist.SFO <= 400000, 1, 0),
                drive.JFK = ifelse(gdist.JFK <= 400000, 1, 0),
                drive.ORD = ifelse(gdist.ORD <= 400000, 1, 0),
                drive.MSY = ifelse(gdist.MSY <= 400000, 1, 0),
                drive.YYZ = ifelse(gdist.YYZ <= 400000, 1, 0),
                drive.LAX = ifelse(gdist.LAX <= 400000, 1, 0),
                drive.DFW = ifelse(gdist.DFW <= 400000, 1, 0),
                drive.DCA = ifelse(gdist.DCA <= 400000, 1, 0),
                drive.YUL = ifelse(gdist.YUL <= 400000, 1, 0),
                drive.MIA = ifelse(gdist.MIA <= 400000, 1, 0),
                drive.SAN = ifelse(gdist.SAN <= 400000, 1, 0),
                drive.ATL = ifelse(gdist.ATL <= 400000, 1, 0),
                drive.HNL = ifelse(gdist.HNL <= 400000, 1, 0)
                )]

final.phl[, `:=`(drive.PHL = ifelse(gdist.PHL <= 400000, 1, 0),
                 drive.SFO = ifelse(gdist.SFO <= 400000, 1, 0),
                 drive.JFK = ifelse(gdist.JFK <= 400000, 1, 0),
                 drive.ORD = ifelse(gdist.ORD <= 400000, 1, 0),
                 drive.MSY = ifelse(gdist.MSY <= 400000, 1, 0),
                 drive.YYZ = ifelse(gdist.YYZ <= 400000, 1, 0),
                 drive.LAX = ifelse(gdist.LAX <= 400000, 1, 0),
                 drive.DFW = ifelse(gdist.DFW <= 400000, 1, 0),
                 drive.DCA = ifelse(gdist.DCA <= 400000, 1, 0),
                 drive.YUL = ifelse(gdist.YUL <= 400000, 1, 0),
                 drive.MIA = ifelse(gdist.MIA <= 400000, 1, 0),
                 drive.SAN = ifelse(gdist.SAN <= 400000, 1, 0),
                 drive.ATL = ifelse(gdist.ATL <= 400000, 1, 0),
                 drive.HNL = ifelse(gdist.HNL <= 400000, 1, 0)
                 )]
```


## Descriptive Statistics
```{r Geodistance Stats}
# Summary Stats for unique location 
summary(final.nyc$gdist.JFK)
summary(final.sf$gdist.SFO)
summary(final.phl$gdist.PHL)
# For some reason there is a single NA values in each set

# Summary Stats for total distances traveled
summary(final.nyc$Frequency * final.nyc$gdist.JFK)
summary(final.sf$Frequency * final.sf$gdist.SFO)
summary(final.phl$Frequency * final.phl$gdist.PHL)

# Total distance traveled in kilometers
total.nyc <- sum(final.nyc$Frequency * final.nyc$gdist.JFK, na.rm = TRUE)/1000
total.sf <- sum(final.sf$Frequency * final.sf$gdist.SFO, na.rm = TRUE)/1000
total.phl <- sum(final.phl$Frequency * final.phl$gdist.PHL, na.rm = TRUE)/1000
total.distance <- rbind(total.nyc, total.sf, total.phl)
```

```{r Drive/Fly}
# Filter for drivers
drive.nyc <- final.nyc[drive.JFK == 1]
drive.sf <- final.sf[drive.SFO == 1]
drive.phl <- final.phl[drive.PHL == 1]

# Filter to flyers
fly.nyc <- final.nyc[drive.JFK != 1]
fly.sf <- final.sf[drive.SFO != 1]
fly.phl <- final.phl[drive.PHL != 1]

# Calculated percentage of drivers
nyc.drive.prop <- sum(drive.nyc$Frequency) / sum(final.nyc$Frequency)
sf.drive.prop <- sum(drive.sf$Frequency) / sum(final.sf$Frequency)
phl.drive.prop <- sum(drive.phl$Frequency) / sum(final.phl$Frequency)
drive.prop <- rbind(nyc.drive.prop, sf.drive.prop, phl.drive.prop)

# Summary Stats for total distances traveled
summary(drive.nyc$Frequency * drive.nyc$gdist.JFK)
summary(drive.sf$Frequency * drive.sf$gdist.SFO)
summary(drive.phl$Frequency * drive.phl$gdist.PHL)
summary(fly.nyc$Frequency * fly.nyc$gdist.JFK)
summary(fly.sf$Frequency * fly.sf$gdist.SFO)
summary(fly.phl$Frequency * fly.phl$gdist.PHL)

# Total one-way distance traveled in kilometers
## Drivers
total.drive.nyc <- sum(drive.nyc$Frequency * drive.nyc$gdist.JFK,
                       na.rm = TRUE)/1000
total.drive.sf <- sum(drive.sf$Frequency * drive.sf$gdist.SFO,
                      na.rm = TRUE)/1000
total.drive.phl <- sum(drive.phl$Frequency * drive.phl$gdist.PHL,
                       na.rm = TRUE)/1000
total.drive <- rbind(total.drive.nyc, total.drive.sf, total.drive.phl)

## Flyers
total.fly.nyc <- sum(fly.nyc$Frequency * fly.nyc$gdist.JFK,
                       na.rm = TRUE)/1000
total.fly.sf <- sum(fly.sf$Frequency * fly.sf$gdist.SFO,
                      na.rm = TRUE)/1000
total.fly.phl <- sum(fly.phl$Frequency * fly.phl$gdist.PHL,
                       na.rm = TRUE)/1000
total.fly <- rbind(total.fly.nyc, total.fly.sf, total.fly.phl)

```


```{r International / Domestic}
# Filter for domestic attendees
dom.nyc <- final.nyc[COUNTRY == "United States"]
dom.sf <- final.sf[COUNTRY == "United States"]
dom.phl <- final.phl[COUNTRY == "United States"]

# International attendees
intl.nyc <- final.nyc[COUNTRY != "United States"]
intl.sf <- final.sf[COUNTRY != "United States"]
intl.phl <- final.phl[COUNTRY != "United States"]

# Summary Counts (% Domestic - More in depth in Excel pivot tables)
nyc.dom.prop <- sum(dom.nyc$Frequency) / sum(final.nyc$Frequency)
sf.dom.prop <- sum(dom.sf$Frequency) / sum(final.sf$Frequency)
phl.dom.prop <- sum(dom.phl$Frequency) / sum(final.phl$Frequency)
dom.prop <- rbind(nyc.dom.prop, sf.dom.prop, phl.dom.prop)

# Summary Stats for total distances traveled
summary(dom.nyc$Frequency * dom.nyc$gdist.JFK)
summary(dom.sf$Frequency * dom.sf$gdist.SFO)
summary(dom.phl$Frequency * dom.phl$gdist.PHL)
summary(intl.nyc$Frequency * intl.nyc$gdist.JFK)
summary(intl.sf$Frequency * intl.sf$gdist.SFO)
summary(intl.phl$Frequency * intl.phl$gdist.PHL)

# Total one-way distance traveled in kilometers
total.dom.nyc <- sum(dom.nyc$Frequency * dom.nyc$gdist.JFK, na.rm = TRUE)/1000
total.dom.sf <- sum(dom.sf$Frequency * dom.sf$gdist.SFO, na.rm = TRUE)/1000
total.dom.phl <- sum(dom.phl$Frequency * dom.phl$gdist.PHL, na.rm = TRUE)/1000
total.domestic <- rbind(total.dom.nyc, total.dom.sf, total.dom.phl)

# Total one-way distance traveled in kilometers
total.intl.nyc <- sum(intl.nyc$Frequency *
                        intl.nyc$gdist.JFK, na.rm = TRUE)/1000
total.intl.sf <- sum(intl.sf$Frequency *
                       intl.sf$gdist.SFO, na.rm = TRUE)/1000
total.intl.phl <- sum(intl.phl$Frequency *
                        intl.phl$gdist.PHL, na.rm = TRUE)/1000
total.international <- rbind(total.intl.nyc, total.intl.sf, total.intl.phl)

# Combine distances into data.fame
table.names <- c("NYC 2018", "SF 2019", "PHL 2020")


total.table <- as.data.frame(cbind(table.names,
                                   total.drive,
                                   total.fly,
                                   drive.prop,
                                   total.domestic,
                                   total.international,
                                   total.distance,
                                   dom.prop))
# Fix headers
colnames(total.table) <- c("Meeting Location",
                           "Distance Driven (km)",
                           "Distance Flown (km)",
                           "Proportion of Drivers (%)",
                           "Domestic Travel (km)",
                           "International Travel (km)",
                           "Total (km)",
                           "Domestic Attendance (%)")
```


## Write Excel Copy
### Intended for subsequent analysis w/o R
```{r Write Geo-Unique Frequncy Data}
final.sets <- list("NYC 2018" = final.nyc,
                 "SF 2019" = final.sf,
                 "PHL 2020" = final.phl,
                 "Convention Centers" = conventions,
                 "Summary Statistics" = total.table)

write.xlsx(final.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Geodistance Data.xlsx")
```
