==> code/05 - Locating Nearest Airport.Rmd <==
---
title: "05 - Locating Nearest Airport"
author: "Alec Stashevsky"
date: "7/31/2020"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(readxl)
library(openxlsx)
library(tidyverse)
library(data.table)
library(geosphere)
rm(list = ls()) # Clean R Environment
```


## Import APA Geodistance Data
```{r APA Import}
path.in <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Geodistance Data.xlsx"

nyc <- read_excel(path.in, sheet = "NYC 2018") %>% setDT()
sf <- read_excel(path.in, sheet = "SF 2019") %>% setDT()
phl <- read_excel(path.in, sheet = "PHL 2020") %>% setDT()
conventions <- read_xlsx(path.in, sheet = "Convention Centers") %>%
  setDT()
```


## Import Global Airport Database 
```{r Airports Import}
path.airports <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GlobalAirportDatabase/GlobalAirportDatabase.txt"

# Import that data
airports.raw <- read.delim(path.airports, header = FALSE, sep = ":") %>% setDT()

# Drop extraneous columns
airports <- airports.raw[, c(1:5, 14:16)]

# Set column names
setnames(airports, names(airports), c("ICAO", "IATA", "Airport Name",
         "City", "Country", "Altitude", "Latitude", "Longitude"))

# Get destination airports
airports.dests <- rbind(airports[IATA == 'PHL'],
                        airports[IATA == 'SFO'],
                        airports[IATA == 'JFK'],
                        airports[IATA == 'ORD'],
                        airports[IATA == 'MSY'],
                        airports[IATA == 'YYZ'],
                        airports[IATA == 'LAX'],
                        airports[IATA == 'DFW'],
                        airports[IATA == 'DCA'],
                        airports[IATA == 'YUL'],
                        airports[IATA == 'MIA'],
                        airports[IATA == 'SAN'],
                        airports[IATA == 'ATL'],
                        airports[IATA == 'HNL']
                        )

# Attach onto the convention center sheet
conventions.airports <- cbind(conventions, airports.dests)
```


## Clean Airport Database
The airport data is assigning very small airports in about 5% of cases. The GoClimate API does not recognize these smaller airport's IATA codes, so we will reassign this list of misqueries to the next closest airports iteratively, until all are read by the API. 

```{r Clean Airports}
# Import the misqueries from first API run
path.misqueries <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/GoClimate API Misqueries.xlsx"

misqueries <- read_excel(path.misqueries, sheet = "Misqueries") %>% setDT()

# It looks like there are quite a few airports w/o coordinates - drop these
# Keep only airports w/ IATA codes
airport.iata <- airports[Latitude != 0 & Longitude != 0 & IATA != 'N/A']

# Remove rows with IATA code in mis-query data
airport.coords <- airport.iata[!(IATA %in% misqueries$IATA)]

```


## Locate nearest aiport 
```{r NYC Airport Location}
# Initialize logs
nyc.airport.log <- rep(NA, nrow(nyc))
nyc.airport.dist <- rep(NA, nrow(nyc))

# Find the closest airport for each NYC attendee origin
for (i in 1:nrow(nyc)){
  nyc.airport.log[i] <- which.min(distGeo(c(nyc$lon [i],
                                            nyc$lat[i]),
                                          airport.coords[, c('Longitude',
                                                             'Latitude')]))
  nyc.airport.dist[i] <- min(distGeo(c(nyc$lon [i],
                                       nyc$lat[i]),
                                     airport.coords[, c('Longitude',
                                                        'Latitude')]))
}

# Get list of attendee airport codes
closest.airports <- airport.coords[nyc.airport.log]

# Merge onto NYC data
nyc.final <- cbind(nyc,
                   closest.airports,
                   airport_dist_meters = nyc.airport.dist)
```

```{r SF Airport Location}
# Initialize logs
sf.airport.log <- rep(NA, nrow(sf))
sf.airport.dist <- rep(NA, nrow(sf))

# Find the closest airport for each sf attendee origin
for (i in 1:nrow(sf)){
  sf.airport.log[i] <- which.min(distGeo(c(sf$lon [i],
                                            sf$lat[i]),
                                          airport.coords[, c('Longitude',
                                                             'Latitude')]))
  sf.airport.dist[i] <- min(distGeo(c(sf$lon [i],
                                       sf$lat[i]),
                                     airport.coords[, c('Longitude',
                                                        'Latitude')]))
}

# Get list of attendee airport codes
closest.airports <- airport.coords[sf.airport.log]

# Merge onto sf data
sf.final <- cbind(sf,
                   closest.airports,
                   airport_dist_meters = sf.airport.dist)
```

```{r PHL Airport Location}
# Initialize logs
phl.airport.log <- rep(NA, nrow(phl))
phl.airport.dist <- rep(NA, nrow(phl))

# Find the closest airport for each phl attendee origin
for (i in 1:nrow(phl)){
  phl.airport.log[i] <- which.min(distGeo(c(phl$lon [i],
                                            phl$lat[i]),
                                          airport.coords[, c('Longitude',
                                                             'Latitude')]))
  phl.airport.dist[i] <- min(distGeo(c(phl$lon [i],
                                       phl$lat[i]),
                                     airport.coords[, c('Longitude',
                                                        'Latitude')]))
}

# Get list of attendee airport codes
closest.airports <- airport.coords[phl.airport.log]

# Merge onto phl data
phl.final <- cbind(phl,
                   closest.airports,
                   airport_dist_meters = phl.airport.dist)
```

## Save Airport Data
```{r Export Data}
final.sets <- list("NYC 2018" = nyc.final,
                 "SF 2019" = sf.final,
                 "PHL 2020" = phl.final,
                 "Convention Centers" = conventions.airports)

write.xlsx(final.sets,
           file = "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/GeoData/APA Geodistance + Airport Data.xlsx")
```
