==> code/08 - Emissions Plots.Rmd <==
---
title: "08 - Emissions Plots"
author: "Alec Stashevsky"
date: "8/25/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(data.table)
library(readxl)
library(openxlsx)
library(ggmap) # Need to cite 
library(maps)
library(rgeos)
library(maptools)
library(geosphere)
library(viridis)
library(gridExtra)
library(grid)
library(plyr)
library(scales)
rm(list = ls()) # Clean R Environment
```

## Import Per Capita Emissions Analysis
```{r Import}
path.emissions <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/APA Emissions Per Capita.xlsx"

path.conventions <- "C:/Users/Alec/Documents/GAP Research/GAP Climate Research/Analysis/APA Emissions All Locations (Analysis).xlsx"

total <- read_excel(path.emissions, sheet = "Emissions Overview") %>% setDT()
drivers <- read_excel(path.emissions, sheet = "Drivers") %>% setDT()
plot.data <- read_excel(path.emissions, sheet = "R Plot Data") %>% setDT()
intl.domestic <- read_excel(path.emissions, sheet = "Domestic-International Totals",
                            range = "A1:F57") %>% setDT()

conventions <- read_xlsx(path.conventions, sheet = "Convention Centers") %>%
  setDT()
```

## Create Bar Plots
```{r Bar Plots}
pdf("APA Emissions Plots.pdf")

# Fix Axis w/ Scientific Notation
options(scipen=10000)

# Set colors
col.nyc <- adjustcolor("orange", alpha = 0.8)
col.sf <- adjustcolor("cyan", alpha = 0.8)
cols <- c(col.nyc, col.sf)

# Set axis coloring
plot.data[, axis.col := c("Actually Held", rep("NYC 2018", 13), rep("SF 2019", 12), "Actually Held", "SF 2019")]

# Create Factor levels for consistent axis
plot.data$Destination <- factor(plot.data$Destination, levels = c("New York City",
                                                                  "Philadelphia",
                                                                  "Washington DC",
                                                                  "Toronto",
                                                                  "Montreal",
                                                                  "Chicago",
                                                                  "Atlanta",
                                                                  "New Orleans",
                                                                  "Dallas",
                                                                  "Miami",
                                                                  "Los Angeles",
                                                                  "San Diego",
                                                                  "San Francisco",
                                                                  "Honolulu")
                                )

# Plot Total Emissions
ggplot(plot.data, aes(fct_rev(fct_reorder(factor(Destination),
                                          # Convert to Metric Tonnes
                                          `Total CO2 Emissions (Kg)`)),
                      `Total CO2 Emissions (Kg)`/1000, 
                      fill = axis.col)) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip() +
  xlab("Conference Destination") +
  ylab("Metric Tonnes CO2") +
  ggtitle("Total CO2 Emissions") +
  guides(fill = guide_legend(title = "Attendance Base", reverse = TRUE)) +
  scale_fill_manual(values = c("Actually Held" = "red",
                               "NYC 2018" = col.nyc,
                               "SF 2019" = col.sf))

# Plot Per Capita Emissions
ggplot(plot.data, aes(fct_rev(fct_reorder(factor(Destination),
                                          # Convert to Metric Tonnes
                                          `CO2e per Capita (Kg)`)),
                      `CO2e per Capita (Kg)`/1000, 
                      fill = axis.col)) +
  geom_bar(stat = "identity", position = "dodge") +
  coord_flip() +
  xlab("Conference Destination") +
  ylab("Metric Tonnes CO2") +
  ggtitle("Emissions per Capita") +
  guides(fill = guide_legend(title = "Attendance Base", reverse = TRUE)) +
  scale_fill_manual(values = c("Actually Held" = "red",
                               "NYC 2018" = col.nyc,
                               "SF 2019" = col.sf))

# Delta Percentage dumbbell Plots
ggplot(plot.data, aes(x = fct_rev(Destination),
                      y = `Percent Delta Increase CO2`,
                      label = round(`Percent Delta Increase CO2`*100))) +
  geom_hline(yintercept = 0) +
  geom_segment(y = 0, 
               x = plot.data$Destination,
               yend = plot.data$`Percent Delta Increase CO2`,
               xend = plot.data$Destination,
               color = 'black',
               size = 1) +
  geom_point(stat = 'identity',
             size = 6,
             color = ifelse(plot.data$`Percent Delta Increase CO2` == 0,
                            "red", "black")) +
  
  geom_text(color="white", size=2.5) +
  facet_wrap(~`Attendance Base`) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 5L), limits = c(-0.5, 1.7)) +
  #ylim(-50, 170) +
  xlab("Conference Destination") +
  ylab("Change from Baseline (red)") +
  ggtitle("Change in CO2e Across Alternate APA Locations") +
  coord_flip()

# Delta Percentage dumbbell Plots (Single Plot)
ggplot(plot.data, aes(x = fct_rev(Destination),
                      `Percent Delta Increase CO2`*100,
                      label = round(`Percent Delta Increase CO2`*100))) +
  geom_hline(yintercept = 0) +
  geom_segment(y = 0, 
               x = plot.data$Destination,
               yend = plot.data$`Percent Delta Increase CO2`*100,
               xend = plot.data$Destination,
               color = 'gray50',
               size = 1) +
  geom_point(stat = 'identity',
             size = 6,
             color = ifelse(plot.data$`Attendance Base` == "SF 2019" &
                              plot.data$`Percent Delta Increase CO2` != 0,
                            col.sf,
                            ifelse(plot.data$`Attendance Base` == "NYC 2018" &
                                     plot.data$`Percent Delta Increase CO2` != 0,
                                   col.nyc,
                                   ifelse(plot.data$`Percent Delta Increase CO2` == 0,
                                          "red", "black")))) +
  geom_text(color="black", size=2.5) +
  ylim(-50, 170) +
  xlab("Conference Destination") +
  ylab("Percentage Change from Baseline (red)") +
  ggtitle("Change in CO2e Across Alternate APA Locations") +
  guides(fill = guide_legend(title = "Attendance Base", reverse = TRUE)) +
  coord_flip()
```

## International / Domestic Breakdown
```{r Domestic/International}
# Create Factor levels for consistent axis
intl.domestic$Destination <- factor(intl.domestic$Destination,
                                    levels = c("New York City",
                                               "Philadelphia",
                                               "Washington DC",
                                               "Toronto",
                                               "Montreal",
                                               "Chicago",
                                               "Atlanta",
                                               "New Orleans",
                                               "Dallas",
                                               "Miami",
                                               "Los Angeles",
                                               "San Diego",
                                               "San Francisco",
                                               "Honolulu")
                                    )


# Total CO2e - No Grouping
ggplot(intl.domestic, aes(x = fct_rev(Destination),
                          y = `Total CO2e (Kg)`/1000)) +
  geom_col(position="stack") +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  ylab("Metric Tons CO2") +
  scale_y_continuous(labels = comma) +
  coord_flip() +
  ggtitle("Total Emissions")


# Change Intl Emissions Shares to Null for Plot labels
intl.domestic.total.e <- intl.domestic
intl.domestic.total.e[`Travel Type` == "International"]$`Emissions Share` <- NA

# Total CO2e
ggplot(intl.domestic.total.e, aes(x = fct_rev(Destination),
                          y = `Total CO2e (Kg)`/1000,
                          fill = `Travel Type`,
                          label = `Emissions Share`)) +
  
  geom_col(position="stack") +
  geom_text(aes(label = percent(`Emissions Share`, accuracy = 5L)),
            size = 2.8,
            position = position_stack(vjust = 0.5)) +
  
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  ylab("Metric Tons CO2") +
  scale_y_continuous(labels = comma, n.breaks = 4) +
  coord_flip() +
  ggtitle("Total Emissions") +
  scale_fill_brewer(palette = "Pastel2") +
  labs(fill="Origin Location")
  


# Total CO2e NYC
ggplot(intl.domestic[`Attendance Base` == "NYC 2018"],
       aes(x = fct_rev(fct_reorder(factor(Destination),
                                   `Total CO2e (Kg)`)),
           y = `Total CO2e (Kg)`/1000,
           fill = `Travel Type`)) +
  
  geom_col(position="stack") +
  xlab("Conference Destination") +
  ylab("Metric Tons CO2") +
  coord_flip() +
  ggtitle("Total Emissions for NYC 2018") +
  scale_fill_brewer(palette = "Pastel2") +
  labs(fill="Origin Location")

# Total CO2e SF
ggplot(intl.domestic[`Attendance Base` == "SF 2019"], aes(x = fct_rev(fct_reorder(factor(Destination),
                                          `Total CO2e (Kg)`)),
                          y = `Total CO2e (Kg)`/1000,
                          fill = `Travel Type`)) +
  geom_col(position="stack") +
  xlab("Conference Destination") +
  ylab("Metric Tons CO2") +
  ylim(c(0, 50000)) + # Match NYC Axis
  coord_flip() +
  ggtitle("Total CO2 Emissions for SF 2019") +
  scale_fill_brewer(palette = "Pastel2") +
  labs(fill="Origin Location")
```


```{r Emissions Share}
# Emissions Share
ggplot(intl.domestic, aes(x = fct_rev(Destination),
                          y = `Emissions Share`,
                          fill = `Travel Type`)) +
  geom_col(position="fill") +
  geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("Share of Total CO2e") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Pastel2") +
  labs(fill="Origin Location")


# Per Capita
ggplot(intl.domestic, aes(x = fct_rev(Destination),
                          y = `Per Capita CO2e (Kg)`/ 1000,
                          fill = `Travel Type`)) +
  geom_col(position = position_dodge2(reverse = TRUE)) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  ylab("Metric Tons CO2") +
  coord_flip() +
  ggtitle("CO2 Emissions per Capita") +
  scale_fill_brewer(palette = "Pastel2") +
  labs(fill="Origin Location")

```


## Transport Mode Plots 
```{r Transport Modes}
# Add driver counts into intl.domestic data
drive.count.plots <- intl.domestic[drivers[`Transport Mode` == "Drive"],
                                   on = c("Destination", "Attendance Base")]

# Keep only domestic counts
drive.count.plots[`Travel Type` == "International"]$Count <- NA

drivers$Destination <- factor(drivers$Destination, levels = c("New York City",
                                                                  "Philadelphia",
                                                                  "Washington DC",
                                                                  "Toronto",
                                                                  "Montreal",
                                                                  "Chicago",
                                                                  "Atlanta",
                                                                  "New Orleans",
                                                                  "Dallas",
                                                                  "Miami",
                                                                  "Los Angeles",
                                                                  "San Diego",
                                                                  "San Francisco",
                                                                  "Honolulu")
                                )



# Emissions Ranking - Kepp w/ consistent factor levels above
nyc.e.ranking <- c("New York City",
                   "Philadelphia",
                   "Washington DC",
                   "Toronto",
                   "Montreal",
                   "Chicago",
                   "Atlanta",
                   "New Orleans",
                   "Dallas",
                   "Miami",
                   "Los Angeles",
                   "San Diego",
                   "San Francisco",
                   "Honolulu")

sf.e.ranking <- c("Washington DC",
                   "Philadelphia",
                   "Chicago",
                   "New York City",
                   "Toronto",
                   "Atlanta",
                   "Montreal",
                   "Dallas",
                   "New Orleans",
                   "Miami",
                   "Los Angeles",
                   "San Diego",
                   "San Francisco",
                   "Honolulu")

# Total CO2e (Both)
ggplot(drivers, 
       aes(x = fct_rev(Destination), # Force Ordering for NYC
           y = `Emissions Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
  #geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("Share of Total CO2e") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")
  


# Total CO2e NYC
ggplot(drivers[`Attendance Base` == "NYC 2018"], 
       aes(x = fct_rev(factor(Destination,
                      levels = nyc.e.ranking)),
           y = `Emissions Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
  #geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("CO2 Emissions NYC 2018 by Transport Mode") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")

# Total CO2e SF
ggplot(drivers[`Attendance Base` == "SF 2019"], 
       aes(x = fct_rev(factor(Destination,
                      levels = sf.e.ranking)),
           y = `Emissions Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
  #geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("CO2 Emissions SF 2019 by Transport Mode") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")
```

```{r Transport Mode Attendance Share}

# Total CO2e (Both)
ggplot(drivers, 
       aes(x = fct_rev(Destination), # Force Ordering for NYC
           y = `Attendance Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
  geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("Conference Attendance") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")

# Total CO2e NYC
ggplot(drivers[`Attendance Base` == "NYC 2018"], 
       aes(x = fct_rev(factor(Destination,
                      levels = nyc.e.ranking)),
           y = `Attendance Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
 # geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("NYC 2018 Conference Attendance") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")

# Total CO2e SF
ggplot(drivers[`Attendance Base` == "SF 2019"], 
       aes(x = fct_rev(factor(Destination,
                      levels = sf.e.ranking)),
           y = `Attendance Share`,
           fill = `Transport Mode`)) +
  geom_col(position="stack") +
 # geom_hline(yintercept = 0.5, lwd = 0.5) +
  facet_wrap(~`Attendance Base`) +
  xlab("Conference Destination") +
  coord_flip() +
  ggtitle("SF 2019 Conference Attendance") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_brewer(palette = "Accent")
```

```{r Export Plots}
dev.off()
```
